move test-pfgemm to attic + minor fixes

+++ tests/testeur_fgemm.C
@@ -28,8 +28,8 @@ using namespace Givaro;
 
 //typedef ModularBalanced<float> Field; 
 //typedef ModularBalanced<double> Field;
-//typedef Givaro::Modular<double> Field;
-typedef Givaro::Modular<float> Field;
+typedef Givaro::Modular<double> Field;
+//typedef Givaro::Modular<float> Field;
 //typedef Givaro::Modular<int> Field; //-> bug avec w>=1 (olddynamic pealing)
 //typedef Givaro::Modular<int32_t> Field;
 //typedef GFqDom<int32_t> Field;
